+++ Data/Implementations/XYZService.cs
@@ -13,116 +13,100 @@ public class XYZService : IXYZService
     {
         _context = context;
     }
+
     public List<XYZ> GetAll()
     {
         return _context.Urls.ToList();
     }
-    
-    public XYZ getById(int id)
+
+    public XYZ? getById(int id)
     {
         var url = _context.Urls.FirstOrDefault(url => url.Id == id);
-        
-        if (url == null)
-        {
-            throw APIException.CreateException(APIException.Code.URL_01, "Url not found",
-                APIException.Type.NOT_FOUND);
-        }
-     
+
         return url;
     }
-    
+
     public XYZ? getUrlLongByShort(string urlShort)
     {
         var url = _context.Urls.FirstOrDefault(url => url.UrlShort == urlShort);
-        
-        if (url == null)
-        {
-            throw APIException.CreateException(APIException.Code.URL_01, "Url not found",
-                APIException.Type.NOT_FOUND);
-        }
-     
+
         return url;
     }
-    
+
     public bool isUrlShortExist(string urlShort)
     {
         if (string.IsNullOrWhiteSpace(urlShort))
         {
-            throw APIException.CreateException(APIException.Code.URL_03, "Url short is required",
-                APIException.Type.BAD_REQUEST);
+            throw new Exception("BD - Url short is required");
         }
-        
+
         return _context.Urls.Any(url => url.UrlShort == urlShort);
     }
-    
+
     public bool isUrlLongExist(string urlLong)
     {
         if (string.IsNullOrWhiteSpace(urlLong))
         {
-            throw APIException.CreateException(APIException.Code.URL_02, "Url long is required",
-                APIException.Type.BAD_REQUEST);
+            throw new Exception("BD - Url long is required");
         }
-        
+
         return _context.Urls.Any(url => url.UrlLong == urlLong);
     }
-    
+
     public XYZ createUrl(XYZForCreationDTO creationDto)
     {
         string randomUrl = UrlMaker.RandomString(6);
-        
+
         while (isUrlShortExist(randomUrl))
         {
             randomUrl = UrlMaker.RandomString(6);
         }
-        
+
         if (!Uri.IsWellFormedUriString(creationDto.UrlLong, UriKind.Absolute))
         {
-            throw APIException.CreateException(APIException.Code.URL_04, "Url long is not valid",
-                APIException.Type.BAD_REQUEST);
+            throw new Exception("BD - Url long is not valid");
         }
-        
+
         var url = new XYZ
         {
             Name = creationDto.Name,
             UrlLong = creationDto.UrlLong,
             UrlShort = randomUrl,
             UserId = creationDto.UserId,
-            CategoryId = _context.Categories.FirstOrDefault(category => category.Name == creationDto.CategoryName.ToLower())?.Id ?? -1,
+            CategoryId = _context.Categories
+                .FirstOrDefault(category => category.Name == creationDto.CategoryName.ToLower())?.Id ?? -1,
         };
-        
-        if (url.CategoryId == -1)
-        {
-            throw APIException.CreateException(APIException.Code.CT_01, "Category not found",
-                APIException.Type.NOT_FOUND);
-        }
-        
+
         try
         {
-            _context.Urls.Add(url); 
+            _context.Urls.Add(url);
         }
-        catch (Exception e)
+        catch (Exception exception)
         {
-            throw APIException.CreateException(APIException.Code.DB_01, e.Message,
-                APIException.Type.INTERNAL_SERVER_ERROR);
+            throw new Exception("IE - An error occurred while setting the data in the database");
         }
-        
+
         try
         {
             _context.SaveChanges();
         }
         catch (Exception e)
         {
-            throw APIException.CreateException(APIException.Code.DB_02, e.Message,
-                APIException.Type.INTERNAL_SERVER_ERROR);
+            throw new Exception("IE - An error occurred while saving the data in the database");
         }
 
         return url;
     }
-    
+
     public void addClick(int id)
     {
         XYZ? urlToChange = getById(id);
-        
+
+        if (urlToChange == null)
+        {
+            throw new Exception("BD - Url not found");
+        }
+
         urlToChange.Clicks++;
 
         try
@@ -131,34 +115,35 @@ public class XYZService : IXYZService
         }
         catch (Exception e)
         {
-            throw APIException.CreateException(APIException.Code.DB_02, e.Message,
-                APIException.Type.INTERNAL_SERVER_ERROR);
+            throw new Exception("IE - An error occurred while saving the data in the database");
         }
-
     }
 
     public void deleteUrl(int id)
     {
         XYZ? urlToDelete = getById(id);
 
+        if (urlToDelete == null)
+        {
+            throw new Exception("BD - Url not found");
+        }
+
         try
         {
             _context.Urls.Remove(urlToDelete);
         }
-        catch (Exception e)
+        catch (Exception exception)
         {
-            throw APIException.CreateException(APIException.Code.DB_01, e.Message,
-                APIException.Type.INTERNAL_SERVER_ERROR);
+            throw new Exception("IE - An error occurred while setting the data in the database");
         }
-        
+
         try
         {
             _context.SaveChanges();
         }
         catch (Exception e)
         {
-            throw APIException.CreateException(APIException.Code.DB_02, e.Message,
-                APIException.Type.INTERNAL_SERVER_ERROR);
+            throw new Exception("IE - An error occurred while saving the data in the database");
         }
     }
 
@@ -166,24 +151,27 @@ public class XYZService : IXYZService
     {
         XYZ? urlToDelete = getUrlLongByShort(urlShort);
 
+        if (urlToDelete == null)
+        {
+            throw new Exception("BD - Url not found");
+        }
+
         try
         {
             _context.Urls.Remove(urlToDelete);
         }
-        catch (Exception e)
+        catch (Exception exception)
         {
-            throw APIException.CreateException(APIException.Code.DB_01, e.Message,
-                APIException.Type.INTERNAL_SERVER_ERROR);
+            throw new Exception("IE - An error occurred while setting the data in the database");
         }
-        
+
         try
         {
             _context.SaveChanges();
         }
         catch (Exception e)
         {
-            throw APIException.CreateException(APIException.Code.DB_02, e.Message,
-                APIException.Type.INTERNAL_SERVER_ERROR);
+            throw new Exception("IE - An error occurred while saving the data in the database");
         }
     }
 }
